requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.10
django-filter==0.13.0
djangorestframework==3.4.4
ipython==5.1.0
ipython-genutils==0.1.0
Markdown==2.6.6
MySQL-python==1.2.5
oauthlib==1.1.2
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.5
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.20
requests==2.11.0
requests-oauthlib==0.6.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
wheel==0.24.0
